+++ Tour Assistant/app/src/main/java/com/example/tourassistant/model/CreateTourResponse.java
@@ -34,16 +34,16 @@ public class CreateTourResponse {
     private Long childs;
     @SerializedName("sourceLat")
     @Expose
-    private Long sourceLat;
+    private double sourceLat;
     @SerializedName("sourceLong")
     @Expose
-    private Long sourceLong;
+    private double sourceLong;
     @SerializedName("desLat")
     @Expose
-    private Long desLat;
+    private double desLat;
     @SerializedName("desLong")
     @Expose
-    private Long desLong;
+    private double desLong;
     @SerializedName("id")
     @Expose
     private Integer id;
@@ -126,15 +126,15 @@ public class CreateTourResponse {
         this.childs = childs;
     }
 
-    public Long getSourceLat() {
+    public double getSourceLat() {
         return sourceLat;
     }
 
-    public void setSourceLat(Long sourceLat) {
+    public void setSourceLat(double sourceLat) {
         this.sourceLat = sourceLat;
     }
 
-    public Long getSourceLong() {
+    public double getSourceLong() {
         return sourceLong;
     }
 
@@ -142,15 +142,15 @@ public class CreateTourResponse {
         this.sourceLong = sourceLong;
     }
 
-    public Long getDesLat() {
+    public double getDesLat() {
         return desLat;
     }
 
-    public void setDesLat(Long desLat) {
+    public void setDesLat(double desLat) {
         this.desLat = desLat;
     }
 
-    public Long getDesLong() {
+    public double getDesLong() {
         return desLong;
     }
